move list to seperate component
